chore: extract utils
